removal of schema tags & usage of int64 type from sequence struct

+++ deployer/internal/controller/openwhisk/main.go
@@ -112,7 +112,7 @@ func greedyControl(obj map[string]interface{}) map[string]interface{} {
 		elapsed = tEnd.Sub(tStart)
 		log.Printf("Elapsed time: %v\n", elapsed)
 		r.PreviousSlack = r.Slack
-		r.Slack += profiledExecutionTimes[i] - elapsed
+		r.Slack += profiledExecutionTimes[i] - int64(elapsed)
 		r.SumOfSlack += r.Slack
 	}
 	log.Printf("Last slack %v\n", r.Slack)
